Build URL Shortener Project setup
I have created a new Spring Boot project for a URL shortener and configured the required Maven dependencies in the pom.xml.

I am using Thymeleaf for the UI in the URL Shortener application. I have designed the layout for the application and configured it in the index.html page.